+++ BitFaster.Caching.UnitTests/Counters/StripedLongAdderTests.cs
@@ -1,5 +1,4 @@
-﻿using System.Threading.Tasks;
-using BitFaster.Caching.Counters;
+﻿using BitFaster.Caching.Counters;
 using FluentAssertions;
 using Xunit;
 
@@ -21,22 +20,6 @@ namespace BitFaster.Caching.UnitTests.Counters
             adder.Increment();
 
             adder.Count().Should().Be(1);
-        }
-
-        [Fact]
-        public async Task WhenAddingConcurrentlySumIsCorrect()
-        {
-            var adder = new Counter();
-
-            await Threaded.Run(4, () => 
-            {
-                for (int i = 0; i < 100_000; i++)
-                {
-                    adder.Increment();
-                }
-            });
-
-            adder.Count().Should().Be(400_000);
         }
     }
 }
