+++ BitFaster.Caching.UnitTests/ScopedCacheTests.cs
@@ -54,22 +54,5 @@ namespace BitFaster.Caching.UnitTests
 
             getOrAdd.Should().Throw<InvalidOperationException>();
         }
-
-        [Fact]
-        public async Task WhenSoakScopedGetOrAddValueIsAlwaysAlive()
-        {
-            for (int i = 0; i < 10; i++)
-            {
-                await Threaded.Run(4, () => {
-                    for (int j = 0; j < 100000; j++)
-                    {
-                        using (var l = this.cache.ScopedGetOrAdd(j, k => new Scoped<Disposable>(new Disposable(k))))
-                        {
-                            l.Value.IsDisposed.Should().BeFalse($"ref count {l.ReferenceCount}");
-                        }
-                    }
-                });
-            }
-        }
     }
 }
